video-api/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
ffmpeg-python==0.2.0
google-cloud-storage==2.9.0 